# spec/api.openapi.yaml
openapi: 3.1.0
info:
  title: Sinjal API
  version: 0.1.0
  description: Small automation/status API for Sinjal.
servers:
  - url: /
components:
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "unauthenticated, forbidden, cross_origin, missing_header, monitor_not_found or internal" }
            message: { type: string }
    Monitor:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        state: { type: string, enum: [up, down, pending, paused, flapping] }
        enabled: { type: boolean }
        state_since: { type: string, format: date-time }
        interval_seconds: { type: integer }
        last_check_at: { type: [string, 'null'], format: date-time }
        last_success_at: { type: [string, 'null'], format: date-time }
        last_failure_at: { type: [string, 'null'], format: date-time }
        tls_not_after: { type: [string, 'null'], format: date-time }
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sinjal_session
  parameters:
    MonitorID:
      in: path
      name: id
      required: true
      schema: { type: string }
    RequestHeader:
      in: header
      name: X-Sinjal-Request
      required: true
      description: Must be "1" on state changes (with the same-origin check, the API's CSRF protection).
      schema: { type: string, enum: ['1'] }
paths:
  /api/v1/status:
    get:
      summary: Instance summary (admin, viewer)
      security: [{ session: [] }]
      responses:
        '200':
          description: Overall status, counts by state and the monitors that need attention
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok, degraded, down] }
                  generated_at: { type: string, format: date-time }
                  monitors: { type: integer }
                  counts: { type: object, additionalProperties: { type: integer } }
                  problems: { type: array, items: { $ref: '#/components/schemas/Monitor' } }
        '401': { description: 'Not signed in', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/monitors:
    get:
      summary: List monitors (admin, viewer)
      security: [{ session: [] }]
      responses:
        '200':
          description: Monitors in name order
          content:
            application/json:
              schema:
                type: object
                properties:
                  monitors: { type: array, items: { $ref: '#/components/schemas/Monitor' } }
        '401': { description: 'Not signed in', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/monitors/{id}:
    get:
      summary: Get monitor (admin, viewer)
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/MonitorID'
      responses:
        '200': { description: Monitor, content: { application/json: { schema: { $ref: '#/components/schemas/Monitor' } } } }
        '401': { description: 'Not signed in', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 'monitor_not_found', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/monitors/{id}/pause:
    post:
      summary: Pause monitor (admin)
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/MonitorID'
        - $ref: '#/components/parameters/RequestHeader'
      responses:
        '204': { description: Paused (also when it already was) }
        '401': { description: 'Not signed in', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'Viewer, cross-origin request or missing header', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 'monitor_not_found', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/monitors/{id}/resume:
    post:
      summary: Resume monitor (admin)
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/MonitorID'
        - $ref: '#/components/parameters/RequestHeader'
      responses:
        '204': { description: Resumed (also when it already was) }
        '401': { description: 'Not signed in', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'Viewer, cross-origin request or missing header', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 'monitor_not_found', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/heartbeat/{token}:
    post:
      summary: Record heartbeat
      description: The body is ignored. GET is accepted too, for clients that cannot POST.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '204': { description: Recorded }
        '404': { description: Invalid token }
        '429': { description: Too many requests from this client address }
  /api/v1/heartbeat:
    post:
      summary: 'Record heartbeat, token in "Authorization: Bearer <token>"'
      responses:
        '204': { description: Recorded }
        '404': { description: Missing or invalid token }
        '429': { description: Too many requests from this client address }
  /status/{slug}/api.json:
    get:
      summary: Public-safe status page JSON
      description: >
        Follows the page's access rules (also at /s/{token}/api.json and, on a
        mapped hostname, /api.json). Components and incidents carry opaque keys,
        never internal ids. A password or authenticated page answers 401 without
        its cookie or a session.
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Public status representation
          content:
            application/json:
              schema:
                type: object
                required: [page, overall_status, generated_at, uptime_90d, components, incidents]
                properties:
                  page:
                    type: object
                    required: [title]
                    properties: { title: { type: string }, description: { type: string } }
                  overall_status:
                    type: object
                    properties: { state: { type: string }, text: { type: string } }
                  generated_at: { type: string, format: date-time }
                  uptime_90d: { type: string }
                  components:
                    type: array
                    items:
                      type: object
                      required: [key, name, status, uptime_90d]
                      properties:
                        key: { type: string }
                        name: { type: string }
                        group: { type: string }
                        status: { type: string }
                        latency_ms: { type: integer }
                        uptime_90d: { type: string }
                  incidents:
                    type: array
                    items:
                      type: object
                      required: [key, component, active, started_at, notes]
                      properties:
                        key: { type: string }
                        component: { type: string, description: Key of the component }
                        active: { type: boolean }
                        started_at: { type: string, format: date-time }
                        ended_at: { type: string, format: date-time }
                        notes:
                          type: array
                          items:
                            type: object
                            properties: { at: { type: string, format: date-time }, message: { type: string } }
        '401': { description: The page needs its password or a session }
        '404': { description: Unknown page }
  /status/{slug}/feed.xml:
    get:
      summary: Incident feed (Atom)
      description: Same access rules as api.json (also at /s/{token}/feed.xml and /feed.xml on a mapped hostname).
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Atom feed, one entry per incident on the page
          content:
            application/atom+xml: {}
        '401': { description: The page needs its password or a session }
        '404': { description: Unknown page }